requirements.txt
mysql-connector-python
pyTelegramBotAPI
tabulate